+++ resources/components/laraknife/buttons/button-position.blade.php
@@ -1,0 +1,13 @@
+@props(['position' => '', 'name' => 'btnOK', 'label' => 'OK', 'width1' => 2, 'width2' => 4])
+@if($position === 'alone' || $position === 'first')
+<div class="row">
+@endif
+@if($width1 > 0)
+<x-laraknife.layout.col-empty width="{{$width1}}"/>
+@endif
+<div class="col-md-{{ $width2 }}">
+    <button class="bg-primary white expand100" name="btnSubmit" value="{{$name}}">{{$label}}</button>
+</div>
+@if($position === 'alone' || $position === 'last')
+</div>
+@endif